requirements.txt
faker
asgiref==3.2.3
certifi==2019.11.28
dj-database-url==0.5.0
Django==3.0.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
wincertstore==0.2
